+++ pkg/controller/zookeepercluster/cfgMap.go
@@ -1,0 +1,46 @@
+package zookeepercluster
+
+import (
+	"context"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+
+	zookeeperv1beta1 "github.com/q8s-io/zookeeper-operator-pravega/pkg/apis/zookeeper/v1beta1"
+	"github.com/q8s-io/zookeeper-operator-pravega/pkg/zk"
+)
+
+func (r *ReconcileZookeeperCluster) reconcileConfigMap(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
+	cm := zk.MakeConfigMap(instance)
+	if err = controllerutil.SetControllerReference(instance, cm, r.scheme); err != nil {
+		return err
+	}
+	foundCm := &corev1.ConfigMap{}
+	err = r.client.Get(context.TODO(), types.NamespacedName{
+		Name:      cm.Name,
+		Namespace: cm.Namespace,
+	}, foundCm)
+	if err != nil && errors.IsNotFound(err) {
+		r.log.Info("Creating a new Zookeeper Config Map",
+			"ConfigMap.Namespace", cm.Namespace,
+			"ConfigMap.Name", cm.Name)
+		err = r.client.Create(context.TODO(), cm)
+		if err != nil {
+			return err
+		}
+		return nil
+	} else if err != nil {
+		return err
+	} else {
+		r.log.Info("Updating existing config-map",
+			"ConfigMap.Namespace", foundCm.Namespace,
+			"ConfigMap.Name", foundCm.Name)
+		zk.SyncConfigMap(foundCm, cm)
+		err = r.client.Update(context.TODO(), foundCm)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
