+++ pkg/controller/zookeepercluster/zookeepercluster_controller.go
@@ -13,29 +13,20 @@ package zookeepercluster
 import (
 	"context"
 	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/go-logr/logr"
 	statefulpodv1 "github.com/q8s-io/iapetos/api/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
 	"sigs.k8s.io/controller-runtime/pkg/source"
-
-	"github.com/q8s-io/zookeeper-operator-pravega/pkg/utils"
-	"github.com/q8s-io/zookeeper-operator-pravega/pkg/yamlexporter"
 
 	zookeeperv1beta1 "github.com/q8s-io/zookeeper-operator-pravega/pkg/apis/zookeeper/v1beta1"
 	"github.com/q8s-io/zookeeper-operator-pravega/pkg/zk"
@@ -66,7 +57,10 @@ func add(mgr manager.Manager, r reconcile.Reconciler) error {
 	if err != nil {
 		return err
 	}
-
+	err = c.Watch(&source.Kind{Type: &statefulpodv1.StatefulPod{}}, &handler.EnqueueRequestForObject{})
+	if err != nil {
+		return err
+	}
 	// Watch for changes to primary resource ZookeeperCluster
 	err = c.Watch(&source.Kind{Type: &zookeeperv1beta1.ZookeeperCluster{}},
 		&handler.EnqueueRequestForObject{})
@@ -118,9 +112,7 @@ type reconcileFun func(cluster *zookeeperv1beta1.ZookeeperCluster) error
 // Reconcile reads that state of the cluster for a ZookeeperCluster object and
 // makes changes based on the state read and what is in the ZookeeperCluster.Spec
 func (r *ReconcileZookeeperCluster) Reconcile(request reconcile.Request) (reconcile.Result, error) {
-	r.log = log.WithValues(
-		"Request.Namespace", request.Namespace,
-		"Request.Name", request.Name)
+	r.log = log.WithValues("Request.Namespace", request.Namespace, "Request.Name", request.Name)
 	r.log.Info("Reconciling ZookeeperCluster")
 
 	// Fetch the ZookeeperCluster instance
@@ -158,119 +150,7 @@ func (r *ReconcileZookeeperCluster) Reconcile(request reconcile.Request) (reconc
 	return reconcile.Result{RequeueAfter: ReconcileTime}, nil
 }
 
-func (r *ReconcileZookeeperCluster) reconcileStatefulPod(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	// we cannot upgrade if cluster is in UpgradeFailed
-	if instance.Status.IsClusterInUpgradeFailedState() {
-		return nil
-	}
-	statusFulPod := zk.MakeStatefulPod(instance)
-	if err = controllerutil.SetControllerReference(instance, statusFulPod, r.scheme); err != nil {
-		return err
-	}
-	foundStatefulPod := &statefulpodv1.StatefulPod{}
-	err = r.client.Get(context.TODO(), types.NamespacedName{
-		Name:      statusFulPod.Name,
-		Namespace: statusFulPod.Namespace,
-	}, foundStatefulPod)
-	if err != nil && errors.IsNotFound(err) {
-		r.log.Info("Creating a new Zookeeper StatefulSet",
-			"StatefulSet.Namespace", statusFulPod.Namespace,
-			"StatefulSet.Name", statusFulPod.Name)
-		return r.client.Create(context.TODO(), statusFulPod)
-
-		return nil
-	} else if err != nil {
-		return err
-	} else {
-		foundSTSSize := *foundStatefulPod.Spec.Size
-		newSTSSize := *statusFulPod.Spec.Size
-		if newSTSSize != foundSTSSize {
-			zkUri := utils.GetZkServiceUri(instance)
-			err = r.zkClient.Connect(zkUri)
-			if err != nil {
-				return fmt.Errorf("Error storing cluster size %v", err)
-			}
-			defer r.zkClient.Close()
-			r.log.Info("Connected to ZK", "ZKURI", zkUri)
-
-			path := utils.GetMetaPath(instance)
-			version, err := r.zkClient.NodeExists(path)
-			if err != nil {
-				return fmt.Errorf("Error doing exists check for znode %s: %v", path, err)
-			}
-
-			data := "CLUSTER_SIZE=" + strconv.Itoa(int(newSTSSize))
-			r.log.Info("Updating Cluster Size.", "New Data:", data, "Version", version)
-			r.zkClient.UpdateNode(path, data, version)
-			return r.updateStatefulSet(instance, foundStatefulPod, statusFulPod)
-		}
-	}
-	return nil
-}
-
-func (r *ReconcileZookeeperCluster) updateStatefulSet(instance *zookeeperv1beta1.ZookeeperCluster,
-	foundSts *statefulpodv1.StatefulPod, statusFulPod *statefulpodv1.StatefulPod) (err error) {
-	r.log.Info("Updating StatefulSet",
-		"StatefulSet.Namespace", foundSts.Namespace,
-		"StatefulSet.Name", foundSts.Name)
-	zk.SyncStatefulSet(foundSts, statusFulPod)
-
-	err = r.client.Update(context.TODO(), foundSts)
-	if err != nil {
-		return err
-	}
-	instance.Status.Replicas = *foundSts.Spec.Size
-	return nil
-}
-
-func (r *ReconcileZookeeperCluster) upgradeStatefulSet(instance *zookeeperv1beta1.ZookeeperCluster, foundSts *statefulpodv1.StatefulPod) (err error) {
-
-	//Getting the upgradeCondition from the zk clustercondition
-	_, upgradeCondition := instance.Status.GetClusterCondition(zookeeperv1beta1.ClusterConditionUpgrading)
-
-	if upgradeCondition == nil {
-		// Initially set upgrading condition to false
-		instance.Status.SetUpgradingConditionFalse()
-		return nil
-	}
-
-	//Setting the upgrade condition to true to trigger the upgrade
-	//When the zk cluster is upgrading Statefulset CurrentRevision and UpdateRevision are not equal and zk cluster image tag is not equal to CurrentVersion
-	if upgradeCondition.Status == corev1.ConditionFalse {
-		if instance.Status.IsClusterInReadyState() /*&& foundSts.Status.CurrentRevision != foundSts.Status.UpdateRevision */ && instance.Spec.Image.Tag != instance.Status.CurrentVersion {
-			instance.Status.TargetVersion = instance.Spec.Image.Tag
-			instance.Status.SetPodsReadyConditionFalse()
-			instance.Status.SetUpgradingConditionTrue("", "")
-		}
-	}
-
-	//checking if the upgrade is in progress
-	if upgradeCondition.Status == corev1.ConditionTrue {
-		//checking when the targetversion is empty
-		if instance.Status.TargetVersion == "" {
-			r.log.Info("upgrading to an unknown version: cancelling upgrade process")
-			return r.clearUpgradeStatus(instance)
-		}
-	}
-	return r.client.Status().Update(context.TODO(), instance)
-}
-
-func (r *ReconcileZookeeperCluster) clearUpgradeStatus(z *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	z.Status.SetUpgradingConditionFalse()
-	z.Status.TargetVersion = ""
-	// need to deep copy the status struct, otherwise it will be overwritten
-	// when updating the CR below
-	status := z.Status.DeepCopy()
-
-	err = r.client.Status().Update(context.TODO(), z)
-	if err != nil {
-		return err
-	}
-
-	z.Status = *status
-	return nil
-}
-
+// for test
 func checkSyncTimeout(z *zookeeperv1beta1.ZookeeperCluster, reason string, updatedReplicas int32, t time.Duration) error {
 	lastCondition := z.Status.GetLastCondition()
 	if lastCondition == nil {
@@ -286,186 +166,4 @@ func checkSyncTimeout(z *zookeeperv1beta1.ZookeeperCluster, reason string, updat
 		}
 	}
 	return nil
-}
-
-func (r *ReconcileZookeeperCluster) reconcileConfigMap(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	cm := zk.MakeConfigMap(instance)
-	if err = controllerutil.SetControllerReference(instance, cm, r.scheme); err != nil {
-		return err
-	}
-	foundCm := &corev1.ConfigMap{}
-	err = r.client.Get(context.TODO(), types.NamespacedName{
-		Name:      cm.Name,
-		Namespace: cm.Namespace,
-	}, foundCm)
-	if err != nil && errors.IsNotFound(err) {
-		r.log.Info("Creating a new Zookeeper Config Map",
-			"ConfigMap.Namespace", cm.Namespace,
-			"ConfigMap.Name", cm.Name)
-		err = r.client.Create(context.TODO(), cm)
-		if err != nil {
-			return err
-		}
-		return nil
-	} else if err != nil {
-		return err
-	} else {
-		r.log.Info("Updating existing config-map",
-			"ConfigMap.Namespace", foundCm.Namespace,
-			"ConfigMap.Name", foundCm.Name)
-		zk.SyncConfigMap(foundCm, cm)
-		err = r.client.Update(context.TODO(), foundCm)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (r *ReconcileZookeeperCluster) reconcileClusterStatus(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	if instance.Status.IsClusterInUpgradingState() || instance.Status.IsClusterInUpgradeFailedState() {
-		return nil
-	}
-	instance.Status.Init()
-	foundPods := &corev1.PodList{}
-	// foundPods := &statefulpodv1.StatefulPodList{}
-	labelSelector := labels.SelectorFromSet(map[string]string{"app": instance.GetName()})
-	listOps := &client.ListOptions{
-		Namespace:     instance.Namespace,
-		LabelSelector: labelSelector,
-	}
-	err = r.client.List(context.TODO(), foundPods, listOps)
-	if err != nil {
-		return err
-	}
-	var (
-		readyMembers   []string
-		unreadyMembers []string
-	)
-	for _, p := range foundPods.Items {
-		ready := true
-		if p.Status.Phase != corev1.PodRunning {
-			ready = false
-		}
-		if ready {
-			readyMembers = append(readyMembers, p.Name)
-		} else {
-			unreadyMembers = append(unreadyMembers, p.Name)
-		}
-	}
-	instance.Status.Members.Ready = readyMembers
-	instance.Status.Members.Unready = unreadyMembers
-	instance.Status.ReadyReplicas = int32(len(foundPods.Items))
-
-	//If Cluster is in a ready state...
-	if instance.Spec.Replicas == instance.Status.ReadyReplicas && (!instance.Status.MetaRootCreated) {
-		r.log.Info("Cluster is Ready, Creating ZK Metadata...")
-		zkUri := utils.GetZkServiceUri(instance)
-		err := r.zkClient.Connect(zkUri)
-		if err != nil {
-			return fmt.Errorf("Error creating cluster metaroot. Connect to zk failed %v", err)
-		}
-		defer r.zkClient.Close()
-		metaPath := utils.GetMetaPath(instance)
-		r.log.Info("Connected to zookeeper:", "ZKUri", zkUri, "Creating Path", metaPath)
-		if err := r.zkClient.CreateNode(instance, metaPath); err != nil {
-			return fmt.Errorf("Error creating cluster metadata path %s, %v", metaPath, err)
-		}
-		r.log.Info("Metadata znode created.")
-		instance.Status.MetaRootCreated = true
-	}
-	r.log.Info("Updating zookeeper status",
-		"StatefulSet.Namespace", instance.Namespace,
-		"StatefulSet.Name", instance.Name)
-	if instance.Status.ReadyReplicas == instance.Spec.Replicas {
-		instance.Status.SetPodsReadyConditionTrue()
-	} else {
-		instance.Status.SetPodsReadyConditionFalse()
-	}
-	if instance.Status.CurrentVersion == "" && instance.Status.IsClusterInReadyState() {
-		instance.Status.CurrentVersion = instance.Spec.Image.Tag
-	}
-	return r.client.Status().Update(context.TODO(), instance)
-}
-
-// YAMLExporterReconciler returns a fake Reconciler which is being used for generating YAML files
-func YAMLExporterReconciler(zookeepercluster *zookeeperv1beta1.ZookeeperCluster) *ReconcileZookeeperCluster {
-	var scheme = scheme.Scheme
-	scheme.AddKnownTypes(zookeeperv1beta1.SchemeGroupVersion, zookeepercluster)
-	return &ReconcileZookeeperCluster{
-		client:   fake.NewFakeClient(zookeepercluster),
-		scheme:   scheme,
-		zkClient: new(zk.DefaultZookeeperClient),
-	}
-}
-
-// GenerateYAML generated secondary resource of ZookeeperCluster resources YAML files
-func (r *ReconcileZookeeperCluster) GenerateYAML(inst *zookeeperv1beta1.ZookeeperCluster) error {
-	if inst.WithDefaults() {
-		fmt.Println("set default values")
-	}
-	for _, fun := range []reconcileFun{
-		r.yamlConfigMap,
-		r.yamlStatefulSet,
-		r.yamlClientService,
-		r.yamlHeadlessService,
-		r.yamlPodDisruptionBudget,
-	} {
-		if err := fun(inst); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// yamlStatefulSet will generates YAML file for StatefulSet
-func (r *ReconcileZookeeperCluster) yamlStatefulSet(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	statefulPod := zk.MakeStatefulPod(instance)
-
-	subdir, err := yamlexporter.CreateOutputSubDir("ZookeeperCluster", statefulPod.Labels["component"])
-	return yamlexporter.GenerateOutputYAMLFile(subdir, statefulPod.Kind, statefulPod)
-}
-
-// yamlClientService will generates YAML file for zookeeper client service
-func (r *ReconcileZookeeperCluster) yamlClientService(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	svc := zk.MakeClientService(instance)
-
-	subdir, err := yamlexporter.CreateOutputSubDir("ZookeeperCluster", "client")
-	if err != nil {
-		return err
-	}
-	return yamlexporter.GenerateOutputYAMLFile(subdir, svc.Kind, svc)
-}
-
-// yamlHeadlessService will generates YAML file for zookeeper headless service
-func (r *ReconcileZookeeperCluster) yamlHeadlessService(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	svc := zk.MakeHeadlessService(instance)
-
-	subdir, err := yamlexporter.CreateOutputSubDir("ZookeeperCluster", "headless")
-	if err != nil {
-		return err
-	}
-	return yamlexporter.GenerateOutputYAMLFile(subdir, svc.Kind, svc)
-}
-
-// yamlPodDisruptionBudget will generates YAML file for zookeeper PDB
-func (r *ReconcileZookeeperCluster) yamlPodDisruptionBudget(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	pdb := zk.MakePodDisruptionBudget(instance)
-
-	subdir, err := yamlexporter.CreateOutputSubDir("ZookeeperCluster", "pdb")
-	if err != nil {
-		return err
-	}
-	return yamlexporter.GenerateOutputYAMLFile(subdir, pdb.Kind, pdb)
-}
-
-// yamlConfigMap will generates YAML file for Zookeeper configmap
-func (r *ReconcileZookeeperCluster) yamlConfigMap(instance *zookeeperv1beta1.ZookeeperCluster) (err error) {
-	cm := zk.MakeConfigMap(instance)
-
-	subdir, err := yamlexporter.CreateOutputSubDir("ZookeeperCluster", "config")
-	if err != nil {
-		return err
-	}
-	return yamlexporter.GenerateOutputYAMLFile(subdir, cm.Kind, cm)
 }
